waveloGPT/requirements.txt
pyproject.toml
docx2txt
aioboto3==11.3.0
aiobotocore
aiohttp==3.8.4
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
atlassian-python-api==3.39.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
boto3
botocore
Brotli==1.0.9
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
ComposeNomadConvertor==0.0.4
ComposeParser==0.0.4
ConfigArgParse==1.5.3
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
dataclasses-json==0.5.8
DateTime==5.2
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.6
dnspython==2.3.0
docutils==0.16
docx2txt==0.8
duckdb==0.8.1
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fastapi==0.98.0
filelock==3.12.2
Flask==2.3.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
gevent
geventhttpclient
gitdb==4.0.10
GitPython==3.1.31
greenlet
h11==0.14.0
numpy
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
identify==2.5.8
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jq==1.6.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.211
langchainplus-sdk==0.0.17
locust==2.11.1
lxml==4.9.2
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.5
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pipenv==2023.7.11
platformdirs==3.8.1
pluggy==1.2.0
posthog==3.0.1
pre-commit==2.20.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
protobuf==4.23.3
psutil==5.9.2
pulsar-client==3.2.0
pyarrow==12.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
pyjson==1.3.0
Pympler==1.0.1
pynput==1.7.6
pyOpenSSL==23.0.0
pyparsing==3.1.0
pypdf==3.16.4
pyreadline3==3.4.1
pyrsistent==0.19.3
pytesseract==0.3.10
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML
redis==4.5.4
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.4.2
roundrobin==0.0.4
rsa==4.7.2
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
starlette==0.27.0
streamlit==1.23.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.19.0
validators==0.20.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.1
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
zope.event==4.5.0
zope.interface==6.1
zstandard==0.21.0
